backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
requests
python-dotenv
